Move datastore into own package as it is logically not connected with the api
