+++ lib/datastore/datastore_test.go
@@ -1,4 +1,4 @@
-package api
+package datastore
 
 import (
 	"testing"
